GT-4: Fixed error when importing qual build.
There was an error about duplicate assessment grading structures
if the structure was disabled, as it wasn't getting loaded properly.

So I added an extra flag to the method that loads them all on the build
to specify whether we want enabled ones only or all.

It's also showing lots of changes on spaces/tabs?