GT-28: Removed all packaged structures to allow for a more 'plugin' style of block.
Pre-built Qual Structures, Levels, Sub Types and Qual Builds will now not be included
in the block.

This is because I cannot keep on top of keeping them up-to-date, so I will instead
keep a database of created structures on the demo site, which the community can contribute to
with structures and quals/units they have created.

This allows me to focus more of my time on the core system, instead of worrying about the
structure data being wrong.